+++ database/migrations/2023_09_24_105126_create_customers_table.php
@@ -33,7 +33,7 @@ return new class extends Migration {
 
         Schema::create('investment', function (Blueprint $table) {
             $table->string('name');
-            $table->decimal('cost', 9, 3);
+            $table->integer('cost');
             $table->primary('name');
         });
     }
